+++ src/main/java/kitchenpos/menu/dto/MenuProductResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.menuproduct.dto;
+package kitchenpos.menu.dto;
 
 import kitchenpos.product.dto.ProductResponse;
